example-python/requirements.txt
streamlit==1.31.1
python-docx==1.1.0

